+++ FindPanel/OgreTextFindResults/NSTextView/OgreTextViewFindResult.m
@@ -21,7 +21,7 @@ static const NSUInteger   OgreTextViewFindResultInitialCapacity = 30;
 
 @implementation OgreTextViewFindResult
 
-- (id)initWithTextView:(NSTextView*)textView
+- (id)initWithTextView:(NSTextView *)textView
 {
 #ifdef DEBUG_OGRE_FIND_PANEL
 	NSLog(@" -initWithString: of %@", [self className]);
@@ -88,7 +88,7 @@ static const NSUInteger   OgreTextViewFindResultInitialCapacity = 30;
 }
 
 /* addMatch */
-- (void)addMatch:(OGRegularExpressionMatch*)match
+- (void)addMatch:(OGRegularExpressionMatch *)match
 {
 	NSRange			range = [match rangeOfMatchedString];
 	NSUInteger		newAbsoluteLocation = range.location;
@@ -124,15 +124,15 @@ static const NSUInteger   OgreTextViewFindResultInitialCapacity = 30;
 	[_matchRangeArray addObject:rangeArray];
 }
 
-- (NSNumber*)lineOfMatchedStringAtIndex:(NSUInteger)index
+- (NSNumber *)lineOfMatchedStringAtIndex:(NSUInteger)index
 {
-    //NSLog(@"lineOfMatchedStringAtIndex:%d", index);
+    //NSLog(@"lineOfMatchedStringAtIndex:%lu", (unsigned long)index);
 	return _lineOfMatchedStrings[(index + 1)];   // 0th is dummy (0番目はダミー)
 }
 
-- (NSAttributedString*)matchedStringAtIndex:(NSUInteger)index
+- (NSAttributedString *)matchedStringAtIndex:(NSUInteger)index
 {
-    //NSLog(@"matchedStringAtIndex:%d", index);
+    //NSLog(@"matchedStringAtIndex:%lu", (unsigned long)index);
 	if (_textView == nil) return [[self textFindResult] missingString];
 	
 	NSArray         *matchArray = _matchRangeArray[(index + 1)];   // 0th is dummy (0番目はダミー)
@@ -217,7 +217,7 @@ static const NSUInteger   OgreTextViewFindResultInitialCapacity = 30;
 @synthesize count = _count;
 
 // [_textView window] will close
-- (void)windowWillClose:(NSNotification*)aNotification
+- (void)windowWillClose:(NSNotification *)aNotification
 {
 #ifdef DEBUG_OGRE_FIND_PANEL
 	NSLog(@"-windowWillClose: of %@", [self className]);
@@ -227,7 +227,7 @@ static const NSUInteger   OgreTextViewFindResultInitialCapacity = 30;
     [[self textFindResult] didUpdate];
 }
 
-- (NSString*)description
+- (NSString *)description
 {
 	return [@{@"Match Line": _lineOfMatchedStrings, 
               @"Match Range": _matchRangeArray,
@@ -235,7 +235,7 @@ static const NSUInteger   OgreTextViewFindResultInitialCapacity = 30;
               @"Count": @(_count)} description];
 }
 
-- (void)textStorageWillProcessEditing:(NSNotification*)aNotification
+- (void)textStorageWillProcessEditing:(NSNotification *)aNotification
 {
 	NSTextStorage   *textStorage = [aNotification object];
 	NSRange			editedRange = [textStorage editedRange];
@@ -243,9 +243,9 @@ static const NSUInteger   OgreTextViewFindResultInitialCapacity = 30;
 	
 	if ([textStorage editedMask] & NSTextStorageEditedCharacters) {
 		// For character of change (文字の変更の場合)
-		/*NSLog(@"w: (%d, %d) -> (%d, %d)", 
-			editedRange.location, editedRange.length - changeInLength, 
-			editedRange.location, editedRange.length);*/
+		/*NSLog(@"w: (%lu, %lu) -> (%lu, %lu)",
+			(unsigned long)editedRange.location, (unsigned long)(editedRange.length - changeInLength), 
+			(unsigned long)editedRange.location, (unsigned long)editedRange.length);*/
 		// Update display of (表示の更新)
 		[self updateOldRange:NSMakeRange(editedRange.location, editedRange.length - changeInLength) newRange:NSMakeRange(editedRange.location, editedRange.length)];
 	}
@@ -313,7 +313,7 @@ static const NSUInteger   OgreTextViewFindResultInitialCapacity = 30;
 	}
 	
 #ifdef DEBUG_OGRE_FIND_PANEL
-	NSLog(@"the maximal index of undisturbed matched string: %d", _updateCacheIndex);
+	NSLog(@"the maximal index of undisturbed matched string: %ld", (long)_updateCacheIndex);
 #endif
 	
 	// Update of absolute for display position cache (表示用絶対位置キャッシュの更新)
@@ -405,7 +405,7 @@ static const NSUInteger   OgreTextViewFindResultInitialCapacity = 30;
     [[self textFindResult] didUpdate];
 }
 
-- (void)updateSubranges:(NSMutableArray*)target count:(NSUInteger)numberOfSubranges oldRange:(NSRange)oldRange newRange:(NSRange)newRange origin:(NSUInteger)origin leftAlign:(BOOL)leftAlign
+- (void)updateSubranges:(NSMutableArray *)target count:(NSUInteger)numberOfSubranges oldRange:(NSRange)oldRange newRange:(NSRange)newRange origin:(NSUInteger)origin leftAlign:(BOOL)leftAlign
 {
 #ifdef DEBUG_OGRE_FIND_PANEL
 	NSLog(@"-updateSubranges: of %@", [self className]);
